#1150: [WIP] cloud extension activation flow
