+++ src/options/pages/marketplace/ConfigureBody.tsx
@@ -20,9 +20,8 @@ import { useField, useFormikContext } from "formik";
 import BootstrapSwitchButton from "bootstrap-switch-button-react";
 import { Card, Table } from "react-bootstrap";
 import { ExtensionPointConfig, RecipeDefinition } from "@/types/definitions";
-import pickBy from "lodash/pickBy";
-import identity from "lodash/identity";
-import { WizardValues } from "@/options/pages/marketplace/wizard";
+import { pickBy, identity } from "lodash";
+import { WizardValues } from "@/options/pages/marketplace/wizardTypes";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCubes, faInfoCircle } from "@fortawesome/free-solid-svg-icons";
 import { Link } from "react-router-dom";
