+++ src/options/pages/marketplace/ServicesBody.tsx
@@ -15,197 +15,23 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { ComponentType, useCallback, useMemo, useState } from "react";
-import ServiceAuthSelector, {
-  AuthOption,
-  useAuthOptions,
-} from "@/options/pages/extensionEditor/ServiceAuthSelector";
+import React, { useMemo } from "react";
+import { useAuthOptions } from "@/options/pages/extensionEditor/ServiceAuthSelector";
 import { uniq } from "lodash";
-import { uuidv4 } from "@/types/helpers";
-import { Button, Card, Table } from "react-bootstrap";
+import { Card, Table } from "react-bootstrap";
 import { Link } from "react-router-dom";
 import { RecipeDefinition, ServiceDefinition } from "@/types/definitions";
 import { useSelectedExtensions } from "@/options/pages/marketplace/ConfigureBody";
-import {
-  faCloud,
-  faInfoCircle,
-  faPlus,
-} from "@fortawesome/free-solid-svg-icons";
+import { faCloud, faInfoCircle } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import ServiceEditorModal from "@/options/pages/services/ServiceEditorModal";
-import { useAsyncState } from "@/hooks/common";
-import registry from "@/services/registry";
-import { RawServiceConfiguration } from "@/core";
-import { servicesSlice } from "@/options/slices";
-import { useDispatch } from "react-redux";
-import { useToasts } from "react-toast-notifications";
-import { useField } from "formik";
-import { persistor } from "@/options/store";
-import { refresh as refreshBackgroundLocator } from "@/background/locator";
-import { GroupTypeBase, MenuListComponentProps } from "react-select";
 import useFetch from "@/hooks/useFetch";
 import { PIXIEBRIX_SERVICE_ID } from "@/services/constants";
-
-const { updateServiceConfig } = servicesSlice.actions;
+import AuthWidget from "@/options/pages/marketplace/AuthWidget";
+import ServiceDescriptor from "@/options/pages/marketplace/ServiceDescriptor";
 
 interface OwnProps {
   blueprint: RecipeDefinition;
 }
-
-const AuthWidget: React.FunctionComponent<{
-  authOptions: AuthOption[];
-  serviceId: string;
-}> = ({ serviceId, authOptions }) => {
-  const helpers = useField(`services.${serviceId}`)[2];
-  const dispatch = useDispatch();
-  const { addToast } = useToasts();
-
-  const [showModal, setShow] = useState(false);
-
-  const [serviceDefinition, isPending, error] = useAsyncState(
-    async () => (await registry.all()).find((x) => x.id === serviceId),
-    [serviceId]
-  );
-
-  const options = useMemo(
-    () => authOptions.filter((x) => x.serviceId === serviceId),
-    [authOptions, serviceId]
-  );
-
-  const save = useCallback(
-    async (values: RawServiceConfiguration) => {
-      const id = uuidv4();
-
-      dispatch(
-        updateServiceConfig({
-          ...values,
-          serviceId,
-          id,
-        })
-      );
-
-      // Need to write the current options to storage so the locator can read them during checks
-      await persistor.flush();
-
-      // Also refresh the service locator on the background so the new auth works immediately
-      await refreshBackgroundLocator({ remote: false, local: true });
-
-      addToast(`Added configuration for integration`, {
-        appearance: "success",
-        autoDismiss: true,
-      });
-
-      // Don't need to track changes locally via setCreated; the new auth automatically flows
-      // through via the redux selectors
-
-      helpers.setValue(id);
-
-      setShow(false);
-    },
-    [helpers, addToast, dispatch, setShow, serviceId]
-  );
-
-  const CustomMenuList = useMemo(() => {
-    const MenuListWithAddButton: ComponentType<
-      MenuListComponentProps<AuthOption, boolean, GroupTypeBase<AuthOption>>
-    > = ({ children }) => (
-      <div>
-        {children}
-        <div className="text-center">
-          <Button
-            size="sm"
-            variant="link"
-            onClick={() => {
-              setShow(true);
-            }}
-          >
-            + Add new
-          </Button>
-        </div>
-      </div>
-    );
-    return MenuListWithAddButton;
-  }, [setShow]);
-
-  const initialConfiguration: RawServiceConfiguration = useMemo(
-    () =>
-      ({
-        serviceId,
-        label: "New Configuration",
-        config: {},
-      } as RawServiceConfiguration),
-    [serviceId]
-  );
-
-  return (
-    <>
-      {showModal && (
-        <ServiceEditorModal
-          configuration={initialConfiguration}
-          service={serviceDefinition}
-          onClose={() => {
-            setShow(false);
-          }}
-          onSave={save}
-        />
-      )}
-
-      <div className="d-inline-flex">
-        {options.length > 0 && (
-          <div style={{ minWidth: "300px" }} className="mr-2">
-            <ServiceAuthSelector
-              name={`services.${serviceId}`}
-              serviceId={serviceId}
-              authOptions={options}
-              CustomMenuList={CustomMenuList}
-            />
-          </div>
-        )}
-        <div>
-          {options.length === 0 && (
-            <Button
-              variant={options.length > 0 ? "info" : "primary"}
-              size="sm"
-              style={{ height: "36px", marginTop: "1px" }}
-              onClick={() => {
-                setShow(true);
-              }}
-              disabled={isPending || error != null}
-            >
-              <FontAwesomeIcon icon={faPlus} /> Configure
-            </Button>
-          )}
-        </div>
-      </div>
-    </>
-  );
-};
-
-const ServiceDescriptor: React.FunctionComponent<{
-  serviceConfigs: ServiceDefinition[];
-  serviceId: string;
-}> = ({ serviceId, serviceConfigs }) => {
-  const config = useMemo(
-    () => serviceConfigs?.find((x) => x.metadata.id === serviceId),
-    [serviceId, serviceConfigs]
-  );
-
-  if (config) {
-    return (
-      <div>
-        <div>{config && <span>{config.metadata.name}</span>}</div>
-        <code className="small p-0">{serviceId}</code>
-      </div>
-    );
-  }
-
-  return (
-    <div>
-      {config && <span>{config.metadata.name}</span>}
-      <code className="p-0">{serviceId}</code>
-    </div>
-  );
-};
 
 const ServicesBody: React.FunctionComponent<OwnProps> = ({ blueprint }) => {
   const [authOptions] = useAuthOptions();
